+++ layer1/Color.c
@@ -125,330 +125,330 @@ void ColorInit(void)
   I->NColor=0;
 
   strcpy(I->Color[I->NColor].Name,"white");
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=1.0;
-  I->Color[I->NColor].Color[2]=1.0;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=1.0F;
+  I->Color[I->NColor].Color[2]=1.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"black");
-  I->Color[I->NColor].Color[0]=0.0;
-  I->Color[I->NColor].Color[1]=0.0;
-  I->Color[I->NColor].Color[2]=0.0;
+  I->Color[I->NColor].Color[0]=0.0F;
+  I->Color[I->NColor].Color[1]=0.0F;
+  I->Color[I->NColor].Color[2]=0.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"blue");
-  I->Color[I->NColor].Color[0]=0.0;
-  I->Color[I->NColor].Color[1]=0.0;
-  I->Color[I->NColor].Color[2]=1.0;
+  I->Color[I->NColor].Color[0]=0.0F;
+  I->Color[I->NColor].Color[1]=0.0F;
+  I->Color[I->NColor].Color[2]=1.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"green");
-  I->Color[I->NColor].Color[0]=0.0;
-  I->Color[I->NColor].Color[1]=1.0;
-  I->Color[I->NColor].Color[2]=0.0;
+  I->Color[I->NColor].Color[0]=0.0F;
+  I->Color[I->NColor].Color[1]=1.0F;
+  I->Color[I->NColor].Color[2]=0.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"red");
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=0.0;
-  I->Color[I->NColor].Color[2]=0.0;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=0.0F;
+  I->Color[I->NColor].Color[2]=0.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"cyan");
-  I->Color[I->NColor].Color[0]=0.0;
-  I->Color[I->NColor].Color[1]=1.0;
-  I->Color[I->NColor].Color[2]=1.0;
+  I->Color[I->NColor].Color[0]=0.0F;
+  I->Color[I->NColor].Color[1]=1.0F;
+  I->Color[I->NColor].Color[2]=1.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"yellow");
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=1.0;
-  I->Color[I->NColor].Color[2]=0.0;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=1.0F;
+  I->Color[I->NColor].Color[2]=0.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"dash");
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=1.0;
-  I->Color[I->NColor].Color[2]=0.0;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=1.0F;
+  I->Color[I->NColor].Color[2]=0.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"violet");
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=0.0;
-  I->Color[I->NColor].Color[2]=1.0;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=0.0F;
+  I->Color[I->NColor].Color[2]=1.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"salmon");
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=0.6;
-  I->Color[I->NColor].Color[2]=0.6;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=0.6F;
+  I->Color[I->NColor].Color[2]=0.6F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"lime");
-  I->Color[I->NColor].Color[0]=0.5;
-  I->Color[I->NColor].Color[1]=1.0;
-  I->Color[I->NColor].Color[2]=0.5;
+  I->Color[I->NColor].Color[0]=0.5F;
+  I->Color[I->NColor].Color[1]=1.0F;
+  I->Color[I->NColor].Color[2]=0.5F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"slate");
-  I->Color[I->NColor].Color[0]=0.5;
-  I->Color[I->NColor].Color[1]=0.5;
-  I->Color[I->NColor].Color[2]=1.0;
+  I->Color[I->NColor].Color[0]=0.5F;
+  I->Color[I->NColor].Color[1]=0.5F;
+  I->Color[I->NColor].Color[2]=1.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"magenta");
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=0.0;
-  I->Color[I->NColor].Color[2]=0.5;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=0.0F;
+  I->Color[I->NColor].Color[2]=0.5F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"orange");
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=0.5;
-  I->Color[I->NColor].Color[2]=0.0;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=0.5F;
+  I->Color[I->NColor].Color[2]=0.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"yellowgreen");
-  I->Color[I->NColor].Color[0]=0.5;
-  I->Color[I->NColor].Color[1]=1.0;
-  I->Color[I->NColor].Color[2]=0.0;
+  I->Color[I->NColor].Color[0]=0.5F;
+  I->Color[I->NColor].Color[1]=1.0F;
+  I->Color[I->NColor].Color[2]=0.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"bluegreen");
-  I->Color[I->NColor].Color[0]=0.0;
-  I->Color[I->NColor].Color[1]=1.0;
-  I->Color[I->NColor].Color[2]=0.5;
+  I->Color[I->NColor].Color[0]=0.0F;
+  I->Color[I->NColor].Color[1]=1.0F;
+  I->Color[I->NColor].Color[2]=0.5F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"blueviolet");
-  I->Color[I->NColor].Color[0]=0.5;
-  I->Color[I->NColor].Color[1]=0.0;
-  I->Color[I->NColor].Color[2]=1.0;
+  I->Color[I->NColor].Color[0]=0.5F;
+  I->Color[I->NColor].Color[1]=0.0F;
+  I->Color[I->NColor].Color[2]=1.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"marine");
-  I->Color[I->NColor].Color[0]=0.0;
-  I->Color[I->NColor].Color[1]=0.5;
-  I->Color[I->NColor].Color[2]=1.0;
+  I->Color[I->NColor].Color[0]=0.0F;
+  I->Color[I->NColor].Color[1]=0.5F;
+  I->Color[I->NColor].Color[2]=1.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"olive");
-  I->Color[I->NColor].Color[0]=0.5;
-  I->Color[I->NColor].Color[1]=0.5;
-  I->Color[I->NColor].Color[2]=0.1;
+  I->Color[I->NColor].Color[0]=0.5F;
+  I->Color[I->NColor].Color[1]=0.5F;
+  I->Color[I->NColor].Color[2]=0.1F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"purple");
-  I->Color[I->NColor].Color[0]=0.7;
-  I->Color[I->NColor].Color[1]=0.2;
-  I->Color[I->NColor].Color[2]=0.7;
+  I->Color[I->NColor].Color[0]=0.7F;
+  I->Color[I->NColor].Color[1]=0.2F;
+  I->Color[I->NColor].Color[2]=0.7F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"teal");
-  I->Color[I->NColor].Color[0]=0.2;
-  I->Color[I->NColor].Color[1]=0.6;
-  I->Color[I->NColor].Color[2]=0.6;
+  I->Color[I->NColor].Color[0]=0.2F;
+  I->Color[I->NColor].Color[1]=0.6F;
+  I->Color[I->NColor].Color[2]=0.6F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"ruby");
-  I->Color[I->NColor].Color[0]=0.5;
-  I->Color[I->NColor].Color[1]=0.1;
-  I->Color[I->NColor].Color[2]=0.1;
+  I->Color[I->NColor].Color[0]=0.5F;
+  I->Color[I->NColor].Color[1]=0.1F;
+  I->Color[I->NColor].Color[2]=0.1F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"forest");
-  I->Color[I->NColor].Color[0]=0.1;
-  I->Color[I->NColor].Color[1]=0.5;
-  I->Color[I->NColor].Color[2]=0.1;
+  I->Color[I->NColor].Color[0]=0.1F;
+  I->Color[I->NColor].Color[1]=0.5F;
+  I->Color[I->NColor].Color[2]=0.1F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"deep");
-  I->Color[I->NColor].Color[0]=0.1;
-  I->Color[I->NColor].Color[1]=0.1;
-  I->Color[I->NColor].Color[2]=0.5;
+  I->Color[I->NColor].Color[0]=0.1F;
+  I->Color[I->NColor].Color[1]=0.1F;
+  I->Color[I->NColor].Color[2]=0.5F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"grey");
-  I->Color[I->NColor].Color[0]=0.5;
-  I->Color[I->NColor].Color[1]=0.5;
-  I->Color[I->NColor].Color[2]=0.5;
+  I->Color[I->NColor].Color[0]=0.5F;
+  I->Color[I->NColor].Color[1]=0.5F;
+  I->Color[I->NColor].Color[2]=0.5F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"gray");
-  I->Color[I->NColor].Color[0]=0.5;
-  I->Color[I->NColor].Color[1]=0.5;
-  I->Color[I->NColor].Color[2]=0.5;
+  I->Color[I->NColor].Color[0]=0.5F;
+  I->Color[I->NColor].Color[1]=0.5F;
+  I->Color[I->NColor].Color[2]=0.5F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"carbon");
-  I->Color[I->NColor].Color[0]=0.2;
-  I->Color[I->NColor].Color[1]=1.0;
-  I->Color[I->NColor].Color[2]=0.2;
+  I->Color[I->NColor].Color[0]=0.2F;
+  I->Color[I->NColor].Color[1]=1.0F;
+  I->Color[I->NColor].Color[2]=0.2F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"nitrogen");
-  I->Color[I->NColor].Color[0]=0.2;
-  I->Color[I->NColor].Color[1]=0.2;
-  I->Color[I->NColor].Color[2]=1.0;
+  I->Color[I->NColor].Color[0]=0.2F;
+  I->Color[I->NColor].Color[1]=0.2F;
+  I->Color[I->NColor].Color[2]=1.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"oxygen");
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=0.3;
-  I->Color[I->NColor].Color[2]=0.3;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=0.3F;
+  I->Color[I->NColor].Color[2]=0.3F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"hydrogen");
-  I->Color[I->NColor].Color[0]=0.9;
-  I->Color[I->NColor].Color[1]=0.9;
-  I->Color[I->NColor].Color[2]=0.9;
+  I->Color[I->NColor].Color[0]=0.9F;
+  I->Color[I->NColor].Color[1]=0.9F;
+  I->Color[I->NColor].Color[2]=0.9F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"sulfer");
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=0.5;
-  I->Color[I->NColor].Color[2]=0.0;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=0.5F;
+  I->Color[I->NColor].Color[2]=0.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"tv_red");
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=0.2;
-  I->Color[I->NColor].Color[2]=0.2;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=0.2F;
+  I->Color[I->NColor].Color[2]=0.2F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"tv_green");
-  I->Color[I->NColor].Color[0]=0.2;
-  I->Color[I->NColor].Color[1]=1.0;
-  I->Color[I->NColor].Color[2]=0.2;
+  I->Color[I->NColor].Color[0]=0.2F;
+  I->Color[I->NColor].Color[1]=1.0F;
+  I->Color[I->NColor].Color[2]=0.2F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"tv_blue");
-  I->Color[I->NColor].Color[0]=0.3;
-  I->Color[I->NColor].Color[1]=0.3;
-  I->Color[I->NColor].Color[2]=1.0;
+  I->Color[I->NColor].Color[0]=0.3F;
+  I->Color[I->NColor].Color[1]=0.3F;
+  I->Color[I->NColor].Color[2]=1.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"tv_yellow");
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=1.0;
-  I->Color[I->NColor].Color[2]=0.2;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=1.0F;
+  I->Color[I->NColor].Color[2]=0.2F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"tv_yellow");
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=1.0;
-  I->Color[I->NColor].Color[2]=0.1;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=1.0F;
+  I->Color[I->NColor].Color[2]=0.1F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"tv_orange");
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=0.55;
-  I->Color[I->NColor].Color[2]=0.15;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=0.55F;
+  I->Color[I->NColor].Color[2]=0.15F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"br0");
-  I->Color[I->NColor].Color[0]=0.1;
-  I->Color[I->NColor].Color[1]=0.1;
-  I->Color[I->NColor].Color[2]=1.0;
+  I->Color[I->NColor].Color[0]=0.1F;
+  I->Color[I->NColor].Color[1]=0.1F;
+  I->Color[I->NColor].Color[2]=1.0F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"br1");
-  I->Color[I->NColor].Color[0]=0.2;
-  I->Color[I->NColor].Color[1]=0.1;
-  I->Color[I->NColor].Color[2]=0.9;
+  I->Color[I->NColor].Color[0]=0.2F;
+  I->Color[I->NColor].Color[1]=0.1F;
+  I->Color[I->NColor].Color[2]=0.9F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"br2");
-  I->Color[I->NColor].Color[0]=0.3;
-  I->Color[I->NColor].Color[1]=0.1;
-  I->Color[I->NColor].Color[2]=0.8;
+  I->Color[I->NColor].Color[0]=0.3F;
+  I->Color[I->NColor].Color[1]=0.1F;
+  I->Color[I->NColor].Color[2]=0.8F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"br3");
-  I->Color[I->NColor].Color[0]=0.4;
-  I->Color[I->NColor].Color[1]=0.1;
-  I->Color[I->NColor].Color[2]=0.7;
+  I->Color[I->NColor].Color[0]=0.4F;
+  I->Color[I->NColor].Color[1]=0.1F;
+  I->Color[I->NColor].Color[2]=0.7F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"br4");
-  I->Color[I->NColor].Color[0]=0.5;
-  I->Color[I->NColor].Color[1]=0.1;
-  I->Color[I->NColor].Color[2]=0.6;
+  I->Color[I->NColor].Color[0]=0.5F;
+  I->Color[I->NColor].Color[1]=0.1F;
+  I->Color[I->NColor].Color[2]=0.6F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"br5");
-  I->Color[I->NColor].Color[0]=0.6;
-  I->Color[I->NColor].Color[1]=0.1;
-  I->Color[I->NColor].Color[2]=0.5;
+  I->Color[I->NColor].Color[0]=0.6F;
+  I->Color[I->NColor].Color[1]=0.1F;
+  I->Color[I->NColor].Color[2]=0.5F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"br6");
-  I->Color[I->NColor].Color[0]=0.7;
-  I->Color[I->NColor].Color[1]=0.1;
-  I->Color[I->NColor].Color[2]=0.4;
+  I->Color[I->NColor].Color[0]=0.7F;
+  I->Color[I->NColor].Color[1]=0.1F;
+  I->Color[I->NColor].Color[2]=0.4F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"br7");
-  I->Color[I->NColor].Color[0]=0.8;
-  I->Color[I->NColor].Color[1]=0.1;
-  I->Color[I->NColor].Color[2]=0.3;
+  I->Color[I->NColor].Color[0]=0.8F;
+  I->Color[I->NColor].Color[1]=0.1F;
+  I->Color[I->NColor].Color[2]=0.3F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"br8");
-  I->Color[I->NColor].Color[0]=0.9;
-  I->Color[I->NColor].Color[1]=0.1;
-  I->Color[I->NColor].Color[2]=0.2;
+  I->Color[I->NColor].Color[0]=0.9F;
+  I->Color[I->NColor].Color[1]=0.1F;
+  I->Color[I->NColor].Color[2]=0.2F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"br9");
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=0.1;
-  I->Color[I->NColor].Color[2]=0.1;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=0.1F;
+  I->Color[I->NColor].Color[2]=0.1F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"pink");
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=0.65;
-  I->Color[I->NColor].Color[2]=0.85;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=0.65F;
+  I->Color[I->NColor].Color[2]=0.85F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"firebrick");
-  I->Color[I->NColor].Color[0]=0.698;
-  I->Color[I->NColor].Color[1]=0.13;
-  I->Color[I->NColor].Color[2]=0.13;
+  I->Color[I->NColor].Color[0]=0.698F;
+  I->Color[I->NColor].Color[1]=0.13F;
+  I->Color[I->NColor].Color[2]=0.13F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"chocolate");
-  I->Color[I->NColor].Color[0]=0.555;
-  I->Color[I->NColor].Color[1]=0.222;
-  I->Color[I->NColor].Color[2]=0.111;
+  I->Color[I->NColor].Color[0]=0.555F;
+  I->Color[I->NColor].Color[1]=0.222F;
+  I->Color[I->NColor].Color[2]=0.111F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"brown");
-  I->Color[I->NColor].Color[0]=0.555;
-  I->Color[I->NColor].Color[1]=0.274;
-  I->Color[I->NColor].Color[2]=0.150;
+  I->Color[I->NColor].Color[0]=0.555F;
+  I->Color[I->NColor].Color[1]=0.274F;
+  I->Color[I->NColor].Color[2]=0.150F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"wheat");
-  I->Color[I->NColor].Color[0]=0.99;
-  I->Color[I->NColor].Color[1]=0.82;
-  I->Color[I->NColor].Color[2]=0.65;
+  I->Color[I->NColor].Color[0]=0.99F;
+  I->Color[I->NColor].Color[1]=0.82F;
+  I->Color[I->NColor].Color[2]=0.65F;
   I->NColor++;
 
   strcpy(I->Color[I->NColor].Name,"grey100"); /* legacy = grey99 */
-  I->Color[I->NColor].Color[0]=1.0;
-  I->Color[I->NColor].Color[1]=1.0;
-  I->Color[I->NColor].Color[2]=1.0;
+  I->Color[I->NColor].Color[0]=1.0F;
+  I->Color[I->NColor].Color[1]=1.0F;
+  I->Color[I->NColor].Color[2]=1.0F;
   I->NColor++;
 
   /* greybow */
 
   for(a=0;a<100;a=a+1) {
     sprintf(I->Color[I->NColor].Name,"grey%02d",a);
-    I->Color[I->NColor].Color[0]=a/99.0;
-    I->Color[I->NColor].Color[1]=a/99.0;
-    I->Color[I->NColor].Color[2]=a/99.0;
+    I->Color[I->NColor].Color[0]=a/99.0F;
+    I->Color[I->NColor].Color[1]=a/99.0F;
+    I->Color[I->NColor].Color[2]=a/99.0F;
     I->NColor++;
   }
 
@@ -457,7 +457,7 @@ void ColorInit(void)
   #define A_DIV 90.9091
 
   for(a=0;a<1000;a=a+1) {
-    set1=(a/A_DIV);
+    set1=(int)(a/A_DIV);
     sprintf(I->Color[I->NColor].Name,"s%03d",a);
     f = 1.0-(a-(set1*A_DIV))/A_DIV;
     I->Color[I->NColor].Color[0]=f*spectrum[set1][0]+(1.0-f)*spectrum[set1+1][0];
